refactor(postgres): modularize postgres driver implementation
Split the 2,867-line postgres.rs into focused submodules (connection,
execution, type_decode, catalog, schema, sql, tests) while keeping the
public PostgresDriver API unchanged.
