+++ packages/drivers/postgres/src/schema.rs
@@ -1,0 +1,217 @@
+//! Schema introspection (columns, table metadata).
+
+use crate::postgres::PostgresDriver;
+use crate::sql::{parse_pg_table_ref, pg_regclass_name};
+use datazen_driver_api::*;
+use sqlx::Row;
+
+impl PostgresDriver {
+    pub(crate) async fn get_columns_impl(
+        &self,
+        handle: &ConnectionHandle,
+        table: &str,
+    ) -> Result<(Vec<ColumnSchema>, Vec<String>), DriverError> {
+        let pools = self.pools.read().await;
+        let pool = Self::get_pool(&pools, handle)?;
+        let (schema, bare_table) = parse_pg_table_ref(table);
+        let regclass = pg_regclass_name(schema, bare_table);
+
+        let cols = sqlx::query(
+            r#"
+                    SELECT column_name, data_type, is_nullable, column_default,
+                           col_description((quote_ident(table_schema)||'.'||quote_ident(table_name))::regclass, ordinal_position) as comment
+                    FROM information_schema.columns
+                    WHERE table_name = $1
+                      AND ($2::text IS NULL OR table_schema = $2)
+                    ORDER BY ordinal_position
+                    "#,
+        )
+        .bind(bare_table)
+        .bind(schema)
+        .fetch_all(pool)
+        .await
+        .map_err(|e| DriverError::QueryFailed(e.to_string()))?;
+
+        // `quote_ident($1)::regclass` fails when the table is not on search_path.
+        // Columns must still load so SQL autocomplete can list fields.
+        let pk_rows = sqlx::query(
+            r#"
+                    SELECT a.attname
+                    FROM pg_index i
+                    JOIN pg_attribute a ON a.attrelid = i.indrelid AND a.attnum = ANY(i.indkey)
+                    WHERE i.indrelid = $1::regclass AND i.indisprimary
+                    "#,
+        )
+        .bind(&regclass)
+        .fetch_all(pool)
+        .await
+        .unwrap_or_default();
+
+        let pk_names: Vec<String> = pk_rows.iter().map(|r| r.get::<String, _>(0)).collect();
+        let columns: Vec<ColumnSchema> = cols
+            .iter()
+            .map(|r| {
+                let name: String = r.get("column_name");
+                let nullable: String = r.get("is_nullable");
+                ColumnSchema {
+                    is_primary_key: pk_names.contains(&name),
+                    name,
+                    data_type: r.get("data_type"),
+                    nullable: nullable == "YES",
+                    default_value: r.get("column_default"),
+                    comment: r.get("comment"),
+                    is_auto_increment: false,
+                }
+            })
+            .collect();
+
+        Ok((columns, pk_names))
+    }
+
+pub(crate) async fn get_table_schema_impl(
+        &self,
+        handle: &ConnectionHandle,
+        table: &str,
+    ) -> Result<TableSchema, DriverError> {
+        let pools = self.pools.read().await;
+        let pool = Self::get_pool(&pools, handle)?;
+        let (schema, bare_table) = parse_pg_table_ref(table);
+        let regclass = pg_regclass_name(schema, bare_table);
+
+        let cols = sqlx::query(
+            r#"
+            SELECT column_name, data_type, is_nullable, column_default,
+                   col_description((quote_ident(table_schema)||'.'||quote_ident(table_name))::regclass, ordinal_position) as comment
+            FROM information_schema.columns
+            WHERE table_name = $1
+              AND ($2::text IS NULL OR table_schema = $2)
+            ORDER BY ordinal_position
+            "#,
+        )
+        .bind(bare_table)
+        .bind(schema)
+        .fetch_all(pool)
+        .await
+        .map_err(|e| DriverError::QueryFailed(e.to_string()))?;
+
+        let pk_rows = sqlx::query(
+            r#"
+            SELECT a.attname
+            FROM pg_index i
+            JOIN pg_attribute a ON a.attrelid = i.indrelid AND a.attnum = ANY(i.indkey)
+            WHERE i.indrelid = $1::regclass AND i.indisprimary
+            "#,
+        )
+        .bind(&regclass)
+        .fetch_all(pool)
+        .await
+        .unwrap_or_default();
+
+        let pk_names: Vec<String> = pk_rows.iter().map(|r| r.get::<String, _>(0)).collect();
+
+        let columns: Vec<ColumnSchema> = cols
+            .iter()
+            .map(|r| {
+                let name: String = r.get("column_name");
+                let nullable: String = r.get("is_nullable");
+                ColumnSchema {
+                    is_primary_key: pk_names.contains(&name),
+                    name,
+                    data_type: r.get("data_type"),
+                    nullable: nullable == "YES",
+                    default_value: r.get("column_default"),
+                    comment: r.get("comment"),
+                    is_auto_increment: false,
+                }
+            })
+            .collect();
+
+        // ── indexes ──
+        let idx_rows = sqlx::query(
+            r#"
+            SELECT i.relname::text                                AS index_name,
+                   array_agg(a.attname::text ORDER BY k.n)        AS columns,
+                   ix.indisunique                                  AS is_unique,
+                   ix.indisprimary                                 AS is_primary,
+                   am.amname::text                                 AS index_type
+            FROM pg_index ix
+            JOIN pg_class i  ON i.oid  = ix.indexrelid
+            JOIN pg_class t  ON t.oid  = ix.indrelid
+            JOIN pg_am   am ON am.oid  = i.relam
+            JOIN LATERAL unnest(ix.indkey) WITH ORDINALITY AS k(attnum, n) ON true
+            JOIN pg_attribute a ON a.attrelid = t.oid AND a.attnum = k.attnum
+            WHERE ix.indrelid = $1::regclass
+            GROUP BY i.relname, ix.indisunique, ix.indisprimary, am.amname
+            ORDER BY ix.indisprimary DESC, i.relname
+            "#,
+        )
+        .bind(&regclass)
+        .fetch_all(pool)
+        .await
+        .unwrap_or_default();
+
+        let indexes: Vec<IndexInfo> = idx_rows
+            .iter()
+            .map(|r| IndexInfo {
+                name: r.get("index_name"),
+                columns: r.get::<Vec<String>, _>("columns"),
+                is_unique: r.get("is_unique"),
+                is_primary: r.get("is_primary"),
+                index_type: r.get("index_type"),
+            })
+            .collect();
+
+        // ── foreign keys ──
+        let fk_rows = sqlx::query(
+            r#"
+            SELECT
+                tc.constraint_name::text                                             AS fk_name,
+                array_agg(kcu.column_name::text ORDER BY kcu.ordinal_position)       AS columns,
+                ccu.table_name::text                                                 AS ref_table,
+                array_agg(ccu.column_name::text ORDER BY kcu.ordinal_position)       AS ref_columns,
+                rc.update_rule::text,
+                rc.delete_rule::text
+            FROM information_schema.table_constraints tc
+            JOIN information_schema.key_column_usage kcu
+              ON kcu.constraint_name = tc.constraint_name
+             AND kcu.table_schema   = tc.table_schema
+            JOIN information_schema.constraint_column_usage ccu
+              ON ccu.constraint_name = tc.constraint_name
+             AND ccu.table_schema   = tc.table_schema
+            JOIN information_schema.referential_constraints rc
+              ON rc.constraint_name = tc.constraint_name
+             AND rc.constraint_schema = tc.table_schema
+            WHERE tc.constraint_type = 'FOREIGN KEY'
+              AND tc.table_name = $1
+              AND ($2::text IS NULL OR tc.table_schema = $2)
+            GROUP BY tc.constraint_name, ccu.table_name, rc.update_rule, rc.delete_rule
+            ORDER BY tc.constraint_name
+            "#,
+        )
+        .bind(bare_table)
+        .bind(schema)
+        .fetch_all(pool)
+        .await
+        .unwrap_or_default();
+
+        let foreign_keys: Vec<ForeignKeyInfo> = fk_rows
+            .iter()
+            .map(|r| ForeignKeyInfo {
+                name: r.get("fk_name"),
+                columns: r.get::<Vec<String>, _>("columns"),
+                referenced_table: r.get("ref_table"),
+                referenced_columns: r.get::<Vec<String>, _>("ref_columns"),
+                on_update: r.get("update_rule"),
+                on_delete: r.get("delete_rule"),
+            })
+            .collect();
+
+        Ok(TableSchema {
+            table_name: table.to_string(),
+            columns,
+            primary_keys: pk_names,
+            indexes,
+            foreign_keys,
+        })
+    }
+}
